removing a bunch of unused code
