+++ app/src/main/java/com/example/clicker/presentation/stream/views/streamManager/ModView.kt
@@ -70,405 +70,24 @@ import androidx.compose.ui.unit.sp
 
 import com.example.clicker.R
 import com.example.clicker.network.clients.BlockedTerm
-import com.example.clicker.network.models.websockets.TwitchUserData
-import com.example.clicker.network.repository.EmoteListMap
-import com.example.clicker.network.repository.EmoteNameUrl
-import com.example.clicker.network.repository.util.AutoModQueueMessage
+
 
 import com.example.clicker.presentation.modView.ListTitleValue
-import com.example.clicker.presentation.modView.ModActionData
-import com.example.clicker.presentation.modView.ModViewDragStateViewModel
-import com.example.clicker.presentation.modView.ModViewViewModel
-import com.example.clicker.presentation.modView.followerModeList
-import com.example.clicker.presentation.modView.slowModeList
-import com.example.clicker.presentation.modView.views.DraggableModViewBox
-import com.example.clicker.presentation.sharedViews.SharedComponents
-import com.example.clicker.presentation.stream.ClickedUIState
-import com.example.clicker.presentation.modView.views.SharedBottomModal
-import com.example.clicker.presentation.modView.views.SmallChat
+
+
 
 import com.example.clicker.presentation.sharedViews.ButtonScope
-import com.example.clicker.presentation.sharedViews.ModViewScaffoldWithDrawer
-import com.example.clicker.presentation.stream.StreamViewModel
-import com.example.clicker.presentation.stream.views.BottomModal
-import com.example.clicker.presentation.stream.views.chat.ChatSettingsColumn
-import com.example.clicker.presentation.stream.views.chat.ChatUI
+
 import com.example.clicker.presentation.stream.views.chat.EmoteOnlySwitch
 import com.example.clicker.presentation.stream.views.chat.FollowersOnlyCheck
-import com.example.clicker.presentation.stream.views.chat.FullChatModView
+
 import com.example.clicker.presentation.stream.views.chat.SlowModeCheck
 import com.example.clicker.presentation.stream.views.chat.SubscriberOnlySwitch
 import com.example.clicker.presentation.stream.views.chat.isScrolledToEnd
-import com.example.clicker.util.Response
-import com.example.clicker.util.WebSocketResponse
-import kotlinx.coroutines.launch
 
 
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun ModViewComponent(
-    closeModView:()->Unit,
-    modViewDragStateViewModel: ModViewDragStateViewModel,
-    inlineContentMap: EmoteListMap,
-    twitchUserChat: List<TwitchUserData>,
-    streamViewModel:StreamViewModel,
-    modViewViewModel:ModViewViewModel,
-    hideSoftKeyboard:()->Unit,
-){
-    val clickedChatterUserState = androidx.compose.material.rememberModalBottomSheetState(
-        initialValue = ModalBottomSheetValue.Hidden,
-        skipHalfExpanded = false
-    )
-    val chatSettingModal = androidx.compose.material.rememberModalBottomSheetState(
-        initialValue = ModalBottomSheetValue.Hidden,
-        skipHalfExpanded = true
-    )
-    val updateClickedUser:(String,String,Boolean,Boolean)->Unit = remember(streamViewModel) { { username, userId, banned, isMod ->
-        streamViewModel.updateClickedChat(
-            username,
-            userId,
-            banned,
-            isMod
-        )
-    } }
-    val doubleClickChat:(String)->Unit =remember(streamViewModel) { {
-        streamViewModel.sendDoubleTapEmote(it)
-    } }
-    val updateMostFrequentEmoteList:(EmoteNameUrl)->Unit =remember(streamViewModel) { {
-        streamViewModel.updateMostFrequentEmoteListTesting(it)
-    } }
-
-    val fullModeActive = modViewDragStateViewModel.fullModeActive.value
-    val scope = rememberCoroutineScope()
-    ModalBottomSheetLayout(
-        sheetBackgroundColor = MaterialTheme.colorScheme.primary,
-        sheetState = clickedChatterUserState,
-        sheetContent = {
-            BottomModal.BottomModalBuilder(
-                clickedUsernameChats = streamViewModel.clickedUsernameChats,
-                clickedUsername = streamViewModel.clickedUIState.value.clickedUsername,
-                bottomModalState = clickedChatterUserState,
-                textFieldValue = streamViewModel.textFieldValue,
-                closeBottomModal = {},
-                banned = streamViewModel.clickedUIState.value.clickedUsernameBanned,
-                unbanUser = {
-                    //  streamViewModel.unBanUser()
-                },
-                //todo: turn this back into --> streamViewModel.state.value.loggedInUserData?.mod ?: false
-                isMod = true,
-                openTimeoutDialog = { streamViewModel.openTimeoutDialog.value = true },
-                openBanDialog = { streamViewModel.openBanDialog.value = true },
-                shouldMonitorUser = streamViewModel.shouldMonitorUser.value,
-                updateShouldMonitorUser = {},
-                clickedUsernameChatsWithDate = streamViewModel.clickedUsernameChatsWithDateSent,
-                openWarnDialog = {}
-            )
-        }
-
-    ){
-        ModViewScaffold(
-            modViewDragStateViewModel=modViewDragStateViewModel,
-            closeModView ={ closeModView() },
-            fullModeActive=fullModeActive,
-            fullChat = {setDraggingFunc->
-                //todo: I need to implement a ModalBottomSheetLayout with the chat settings
-                ModalBottomSheetLayout(
-                    sheetState = chatSettingModal,
-                    sheetContent ={
-                        ChatSettingsColumn(
-                            advancedChatSettings = streamViewModel.advancedChatSettingsState.value,
-                            changeAdvancedChatSettings = {newValue -> streamViewModel.updateAdvancedChatSettings(newValue)},
-                            changeNoChatMode = {newValue -> streamViewModel.setNoChatMode(newValue)},
-
-                            chatSettingsEnabled = streamViewModel.state.value.loggedInUserData?.mod ?: false,
-                            followerModeList= followerModeList,
-                            selectedFollowersModeItem=modViewViewModel.uiState.value.selectedFollowerMode,
-                            changeSelectedFollowersModeItem ={newValue -> modViewViewModel.changeSelectedFollowersModeItem(newValue)},
-                            slowModeList= slowModeList,
-                            selectedSlowModeItem=modViewViewModel.uiState.value.selectedSlowMode,
-                            changeSelectedSlowModeItem ={newValue ->modViewViewModel.changeSelectedSlowModeItem(newValue)},
-                            emoteOnly = modViewViewModel.uiState.value.emoteOnly,
-                            setEmoteOnly = {newValue ->modViewViewModel.updateEmoteOnly(newValue)},
-                            subscriberOnly =modViewViewModel.uiState.value.subscriberOnly,
-                            setSubscriberOnly={newValue -> modViewViewModel.updateSubscriberOnly(newValue)},
-                        )
-                    }
-                ) {
-                    //todo: this is where FullChatModView goes
-                    FullChatModView(
-                        twitchUserChat = twitchUserChat,
-                        showBottomModal={
-                            scope.launch {
-                                clickedChatterUserState.show()
-                            }
-                        },
-                        updateClickedUser = { username, userId, banned, isMod ->
-                            updateClickedUser(
-                                username,
-                                userId,
-                                banned,
-                                isMod
-                            )
-                        },
-                        showTimeoutDialog={
-                            streamViewModel.openTimeoutDialog.value = true
-                        },
-                        showBanDialog = {streamViewModel.openBanDialog.value = true},
-                        doubleClickMessage={ username->
-                            doubleClickChat(username)
-
-                        },
-
-                        showOuterBottomModalState = {
-                            Log.d("BottomModalClicked","showOuterBottomModalState Clicked")
-                            //seems fine that it is empy
-                            scope.launch {
-
-                            }
-                        },
-                        newFilterMethod={newTextValue -> streamViewModel.newParsingAgain(newTextValue)},
-
-                        orientationIsVertical =true,
-
-                        //todo:change back to --> streamViewModel.state.value.loggedInUserData?.mod ?: false
-                        isMod = true,
-                        clickedAutoCompleteText = { username ->
-                            streamViewModel.autoTextChange(username)
-                        },
-                        showModal = {
-                            //todo: This is what is clicked when I want to launch the bottom modal
-                            scope.launch {
-                                chatSettingModal.show()
-                            }
-                        },
-                        notificationAmount =0,
-                        textFieldValue = streamViewModel.textFieldValue,
-                        sendMessageToWebSocket = { string ->
-                            streamViewModel.sendMessage(string)
-                        },
-                        noChat = streamViewModel.advancedChatSettingsState.value.noChatMode,
-                        deleteChatMessage={messageId ->streamViewModel.deleteChatMessage(messageId)},
-                        clickedCommandAutoCompleteText={clickedValue -> streamViewModel.clickedCommandAutoCompleteText(clickedValue)},
-                        inlineContentMap = streamViewModel.inlineTextContentTest.value,
-                        hideSoftKeyboard={
-                            hideSoftKeyboard()
-
-                        },
-                        emoteBoardGlobalList = streamViewModel.globalEmoteUrlList.value,
-                        updateTextWithEmote = {newValue -> streamViewModel.addEmoteToText(newValue)},
-                        emoteBoardChannelList =streamViewModel.channelEmoteUrlList.value,
-                        deleteEmote={streamViewModel.deleteEmote()},
-                        showModView={
-                            closeModView()
-                        },
-                        fullMode = fullModeActive,
-                        setDragging = {
-                            Log.d("doubleClickingThings","CLICKED")
-                            setDraggingFunc()
-                        },
-                        emoteBoardMostFrequentList= streamViewModel.mostFrequentEmoteListTesting.value,
-                        updateMostFrequentEmoteList={value ->updateMostFrequentEmoteList(value)},
-                        globalBetterTTVEmotes=streamViewModel.globalBetterTTVEmotes.value,
-                        channelBetterTTVResponse = streamViewModel.channelBetterTTVEmote.value,
-                        sharedBetterTTVResponse= streamViewModel.sharedChannelBetterTTVEmote.value,
-                        userIsSub = streamViewModel.state.value.loggedInUserData?.sub ?: false,
-                        forwardSlashes = streamViewModel.forwardSlashCommandImmutable.value,
-                        filteredChatListImmutable = streamViewModel.filteredChatListImmutable.value
-                    )
-                }
-
-            },
-            modActionsList = modViewViewModel.modActionsList,
-            modActionStatus = modViewViewModel.modViewStatus.value.modActions,
-            smallChat = {setDraggingFunc->
-                SmallChat(
-                    twitchUserChat=twitchUserChat,
-                    showBottomModal ={
-                        scope.launch {
-                            clickedChatterUserState.show()
-                        }
-                    },
-                    updateClickedUser = { username, userId, banned, isMod ->
-                        updateClickedUser(
-                            username,
-                            userId,
-                            banned,
-                            isMod
-                        )
-                    },
-                    showTimeoutDialog={
-                        streamViewModel.openTimeoutDialog.value = true
-                    },
-                    showBanDialog = {streamViewModel.openBanDialog.value = true},
-                    doubleClickMessage={ username->
-                        doubleClickChat(username)
-                    },
-                    deleteChatMessage={messageId ->streamViewModel.deleteChatMessage(messageId)},
-                    //todo:change back to --> true for testing
-                    isMod =streamViewModel.state.value.loggedInUserData?.mod ?: false,
-                    inlineContentMap =inlineContentMap,
-                    setDragging = {value -> setDraggingFunc()},
-                )
-            },
-            autoModMessageList = modViewViewModel.autoModMessageList,
-            manageAutoModMessage ={
-                    messageId,action -> modViewViewModel.manageAutoModMessage(messageId,action)
-                                  },
-            autoModQueueChecked = modViewViewModel.uiState.value.autoModMessagesNotifications,
-            changeAutoModQueueChecked ={value ->modViewViewModel.changeAutoModQueueChecked(value)},
-            modActionsChecked=modViewViewModel.uiState.value.modActionNotifications,
-            changeModActionsChecked ={value ->modViewViewModel.changeModActionsChecked(value)},
-            autoModStatus = modViewViewModel.modViewStatus.value.autoModMessageStatus
-        )
-
-    }
-
-}
 
 
-// Todo: this is what I am going to try and replace
-    @Composable
-    fun ModViewScaffold(
-        closeModView:()->Unit,
-        modViewDragStateViewModel: ModViewDragStateViewModel,
-
-        modActionStatus: WebSocketResponse<Boolean>,
-        modActionsList: List<ModActionData>,
-        autoModMessageList: List<AutoModQueueMessage>,
-
-        fullModeActive:Boolean,
-        manageAutoModMessage:(String,String)-> Unit,
-        autoModQueueChecked:Boolean,
-        changeAutoModQueueChecked:(Boolean)->Unit,
-
-        modActionsChecked:Boolean,
-        changeModActionsChecked:(Boolean)->Unit,
-        autoModStatus: WebSocketResponse<Boolean>,
-        fullChat: @Composable ( setDraggingTrue: () -> Unit)-> Unit,
-        smallChat: @Composable ( setDraggingTrue: () -> Unit)-> Unit
-        ){
-
-        val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
-        val scope = rememberCoroutineScope()
-
-
-        ModViewScaffoldWithDrawer(
-            topBar = {showDrawerFunc ->
-                Row(modifier = Modifier
-                    .fillMaxWidth()
-                    .background(MaterialTheme.colorScheme.primary)
-                    .padding(10.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween
-                ){
-                    Row(
-
-                    ){
-                        Icon(
-                            imageVector = Icons.Default.Menu,
-                            contentDescription = "close mod view",
-                            modifier = Modifier
-                                .clickable {
-                                    showDrawerFunc()
-                                }
-                                .size(35.dp),
-                            tint = MaterialTheme.colorScheme.onPrimary
-                        )
-                        Text("Mod View", fontSize = MaterialTheme.typography.headlineLarge.fontSize, color = MaterialTheme.colorScheme.onPrimary)
-                    }
-
-                    Icon(
-                        imageVector = Icons.Default.Close,
-                        contentDescription = "close mod view",
-                        modifier = Modifier
-                            .clickable {
-                                closeModView()
-                            }
-                            .size(35.dp),
-                        tint = MaterialTheme.colorScheme.onPrimary
-                    )
-
-
-                }
-            },
-            drawerState=drawerState,
-            bottomBar = {
-                //there is no bottom bar intentionally
-            },
-            showError = modViewDragStateViewModel.showDrawerError.value,
-            checkIndexAvailability ={index ->modViewDragStateViewModel.checkBoxIndexAvailability(index)},
-            autoModQueueChecked = autoModQueueChecked,
-            changeAutoModQueueChecked ={value ->changeAutoModQueueChecked(value)},
-            modActionsChecked=modActionsChecked,
-            changeModActionsChecked ={value ->changeModActionsChecked(value)}
-        ) {contentPadding ->
-            Log.d("contentPaddingModView","contentPadding --> ${contentPadding.calculateTopPadding().value}")
-
-            DraggableModViewBox(
-                contentPaddingValues = contentPadding,
-                boxOneOffsetY =modViewDragStateViewModel.dragStateOffsets.value.boxOneOffsetY,
-                setBoxOneOffset={newValue-> modViewDragStateViewModel.setBoxOneOffset(newValue)},
-                boxOneDragState = modViewDragStateViewModel.boxOneDragState,
-                boxOneZIndex = modViewDragStateViewModel.boxIndexes.value.boxOneZIndex,
-                animateToOnDragStop = modViewDragStateViewModel.animateToOnDragStop,
-                indivBoxSize = modViewDragStateViewModel.indivBoxSize,
-                sectionBreakPoint = modViewDragStateViewModel.sectionBreakPoint,
-
-                boxTwoOffsetY = modViewDragStateViewModel.dragStateOffsets.value.boxTwoOffsetY,
-                boxTwoZIndex = modViewDragStateViewModel.boxIndexes.value.boxTwoZIndex,
-                setBoxTwoOffset = {newValue ->modViewDragStateViewModel.setBoxTwoOffset(newValue)},
-                boxTwoDragState = modViewDragStateViewModel.boxTwoDragState,
-
-                boxThreeZIndex = modViewDragStateViewModel.boxIndexes.value.boxThreeZIndex,
-                boxThreeOffsetY = modViewDragStateViewModel.dragStateOffsets.value.boxThreeOffsetY,
-                setBoxThreeOffset = {newValue ->modViewDragStateViewModel.setBoxThreeOffset(newValue)},
-                boxThreeDragState = modViewDragStateViewModel.boxThreeDragState,
-
-                boxOneDragging = modViewDragStateViewModel.isDragging.value.boxOneDragging,
-                setBoxOneDragging = {newValue -> modViewDragStateViewModel.setBoxOneDragging(newValue)},
-
-                boxThreeDragging =modViewDragStateViewModel.isDragging.value.boxThreeDragging,
-                boxTwoDragging =modViewDragStateViewModel.isDragging.value.boxTwoDragging,
-                setBoxThreeDragging ={
-                        newValue -> modViewDragStateViewModel.setBoxThreeDragging(newValue)
-                    Log.d("WHERETHEDOUBLEIS","DraggableModViewBox")
-                                     },
-                setBoxTwoDragging ={newValue -> modViewDragStateViewModel.setBoxTwoDragging(newValue)},
-
-                deleteOffsetY = modViewDragStateViewModel.deleteOffset.value,
-                boxOneIndex = modViewDragStateViewModel.boxTypeIndex.value.boxOneIndex,
-                boxTwoIndex = modViewDragStateViewModel.boxTypeIndex.value.boxTwoIndex,
-                boxThreeIndex = modViewDragStateViewModel.boxTypeIndex.value.boxThreeIndex,
-                setBoxIndex = {box,value -> modViewDragStateViewModel.changeBoxTypeIndex(box,value)},
-
-                boxOneHeight = modViewDragStateViewModel.indivBoxHeight.value.boxOne,
-                boxTwoHeight = modViewDragStateViewModel.indivBoxHeight.value.boxTwo,
-                boxThreeHeight = modViewDragStateViewModel.indivBoxHeight.value.boxThree,
-
-                modActionStatus=modActionStatus,
-                modActionsList=modActionsList,
-
-                fullModeActive=fullModeActive,
-                fullChat={ setDraggingFunc ->
-                    fullChat(
-                        setDraggingTrue={setDraggingFunc()}
-                    )
-                },
-                smallChat={ setDraggingFunc ->
-                    smallChat(
-                        setDraggingTrue={setDraggingFunc()}
-                    )
-                },
-                autoModMessageList=autoModMessageList,
-                manageAutoModMessage ={messageId,action ->manageAutoModMessage(messageId,action)},
-                autoModStatus=autoModStatus
-
-
-            )
-
-
-        }
-
-    }
 
     /**
      * SectionHeaderRow is a [Row] composable meant to show a [Text] containing [title] and a [ModesHeaderRow] side by side
